pyproject.toml
[tool.poetry]
name = "avalancheengine"
version = "0.1.0"
description = ""
authors = ["DarcJC <me@darc.pro>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.1"
typer = "^0.12.5"
libclang = "^18.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
